add  orientation attr for doc view

+++ library/src/main/java/com/cherry/lib/doc/office/pg/control/PGPrintMode.java
@@ -488,7 +488,7 @@ public class PGPrintMode extends FrameLayout implements IPageListViewListener
      * page list view moving position
      * @param position horizontal or vertical
      */
-    public byte getPageListViewMovingPosition()
+    public int getPageListViewMovingPosition()
     {
     	return control.getMainFrame().getPageListViewMovingPosition();
     }
